Iterate over wanted key
We have a list of keys that correspond to the data we want to dump into our CSV. So far, we can iterate over that list, check whether the key exists in our dictionary parsed from the JSON that the API gives us, and if it exists, it prints it out. This is an intermediate step for our final goal.